Looks like we're good to begin!
Yes, I made an entire library for this project.
I say this with full knowledge of how bad this is because I really
just should've been working on making an MVP but the lack of proper
types was pissing me off. Anyways, I think ts-steam-webapi is a
good example of how to make a real typescript library that's not
bad to use. People should get into the habit of using more newtypes
like I do.

Here's the plan:
1. take snapshots of the world and store them in the db.
2. diff snapshots and create more tables for these diffs.
3. skip the console interface entirely and install electron.
   make a really basic, and I mean two buttons basic, GUI interface
   for this and have it display results in <li>s or something.
4. scan open processes and check if they're running a steam game.
   this may involve having to read Valve's stupid .vdfs in order to
   find out where the users's library is. for now, just assume that
   they're playing a steam game if the path has "steam" in it.
5. integrate into whatever cron-like system is on windows so that
   snapshots are taken monthly and diffs are taken nightly.
6. actually make better GUI. maybe think about nodegui?

For now, I'm going to take a break from programming. Maybe for a
week or more because I have to do schoolwork and I could also use a
bit of relaxation.